Event Sweep Check; Cleanup Auto Register - Added an event sweep to delete old events that are no longer valid - Convereted eventIDs to Long storage rather than String storage to avoid excessive conversion during app usage - Removed leftover autoregister code from webviews

+++ composeApp/src/commonMain/kotlin/org/swg/swiftgroups_app/CGAPI/UpcomingEvents/UpcomingEventData.kt
@@ -18,7 +18,7 @@ data class UpcomingEventData(
     val event_group: String,
     val event_group_id: Int,
     val event_header: String?,
-    val event_id: Int,
+    val event_id: Long,
     val event_name: String,
     val event_start_time: String,
     val event_start_utc: String,
